STY/ENH: Moved functions to trace
