+++ setup.py
@@ -33,7 +33,7 @@ setup(name='OMMBV',
       version=version,
       license='BSD-3-Clause',
       packages=['OMMBV', 'OMMBV.heritage', 'OMMBV.satellite', 'OMMBV.trans',
-                'OMMBV.utils', 'OMMBV.vector'],
+                'OMMBV.trace', 'OMMBV.utils', 'OMMBV.vector'],
       description=' '.join(('Orthogonal geomagnetic vector basis and',
                             'field-line mapping for multipole magnetic',
                             'fields.')),
